Added custom SlugGenerator class
Currently CodiMD supports three per-note configurable modes how to generate slug URLs out of headings. This commit adds the methods necessary for the generation of these.

Be careful: This approach isn't working correctly as in GFM mode a higher number will be appended with each (whitespace) change of the heading.